fix format in notifier settings

+++ lib/memeping/notifications/criteria.ex
@@ -28,8 +28,8 @@ defmodule MemePing.Notifications.Criteria do
   @primary_key false
   embedded_schema do
     for {metric, _label} <- @metrics do
-      field :"#{metric}_min", :integer
-      field :"#{metric}_max", :integer
+      field :"#{metric}_min", :float
+      field :"#{metric}_max", :float
     end
   end
